refactor: Split the components folder into app and shopcart
